Update code after header files change
