+++ teeio-validator/tools/CMakeLists.txt
@@ -20,26 +20,22 @@ INCLUDE_DIRECTORIES(${SPDM_DEVICE_DIR}/include
 SET(src_lside
     lside.c
     ide_common.c
-    ${TEEIO_VALIDATOR_LIB_DIR}/pcie_ide_test_lib/pci_ide.c
-    ${TEEIO_VALIDATOR_LIB_DIR}/pcie_ide_test_lib/scan_pci.c
-    ${TEEIO_VALIDATOR_DIR}/ide_test_ini.c
-    ${TEEIO_VALIDATOR_LIB_DIR}/pcie_ide_test_lib/rp_pci.c)
+    ${TEEIO_VALIDATOR_DIR}/ide_test_ini.c)
 
 SET(lside_LIBRARY
     debuglib
-    helperlib)
+    helperlib
+    pcie_ide_lib)
 
 SET(src_setide
     setide.c
     ide_common.c
-    ${TEEIO_VALIDATOR_LIB_DIR}/pcie_ide_test_lib/pci_ide.c
-    ${TEEIO_VALIDATOR_LIB_DIR}/pcie_ide_test_lib/scan_pci.c
-    ${TEEIO_VALIDATOR_DIR}/ide_test_ini.c
-    ${TEEIO_VALIDATOR_LIB_DIR}/pcie_ide_test_lib/rp_pci.c)
+    ${TEEIO_VALIDATOR_DIR}/ide_test_ini.c)
 
 SET(setide_LIBRARY
     debuglib
-    helperlib)
+    helperlib
+    pcie_ide_lib)
 
 ADD_EXECUTABLE(lside ${src_lside})
 TARGET_LINK_LIBRARIES(lside ${lside_LIBRARY})
